BOT/src/bot_app/bot.py
import os
from dotenv import load_dotenv
from aiogram import Bot, Dispatcher
from aiogram.contrib.fsm_storage.memory import MemoryStorage

load_dotenv()

# Отримуємо токен бота з оточення
TOKEN = os.getenv('BOT_TOKEN')

# Створюємо екземпляр бота та диспетчера
bot = Bot(token=TOKEN)
storage = MemoryStorage()
dp = Dispatcher(bot, storage=storage)
