requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiomysql==0.1.1
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
kombu==5.2.4
magic-filter==1.0.9
multidict==6.0.4
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pydantic==1.10.8
PyMySQL==1.0.3
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
